refactor: Merge kustosql and kustokql in the single kusto module
